d3d render target fixed with depth stencil buffer. If depth resolution is smaller (created for window) than resolution of render target, d3d couldn't render just clear the render target. gl & openvr. Required extension: WGL_NV_DX_interop (dx11 -> WGL_NV_DX_interop2) -> not supported by intel.

+++ VRTestApp/CommonRenderer.h
@@ -170,7 +170,7 @@ protected:
 #endif
 		, mElemSize(ElemSize)
 	{
-		const int len = strlen(ParamName) + 1;
+		const size_t len = strlen(ParamName) + 1;
 		mParamName = new char[len];
 		strcpy_s(mParamName, len, ParamName);
 	}
@@ -312,7 +312,7 @@ public:
 		}
 	}
 
-	int GetElemCount()
+	size_t GetElemCount()
 	{
 		return layout.size();
 	}
@@ -344,7 +344,7 @@ public:
 		return slotInfo[slot].firstElemIdx;
 	}
 
-	unsigned int GetSlotCount()
+	size_t GetSlotCount()
 	{
 		return slotInfo.size();
 	}
